WIP: Revert back to using uint64 for nonce as rlp crate doesn't work with uint256; add ERC20App contract to whitelist

+++ ethereum/contracts/OutboundChannel.sol
@@ -10,7 +10,7 @@ abstract contract OutboundChannel {
 
     event Message(
         address source,
-        uint nonce,
+        uint64 nonce,
         bytes payload
     );
